# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/books": {
      "get": {
        "operationId": "get_books",
        "responses": {
          "200": {
            "description": "List all books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_book",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "operationId": "get_book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Book id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get book by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "operationId": "update_book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Book id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Book updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "operationId": "delete_book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Book id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    },
    "/publishers": {
      "get": {
        "operationId": "get_publishers",
        "responses": {
          "200": {
            "description": "List all publishers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublisherResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_publisher",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublisherCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Publisher created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/publishers/{id}": {
      "get": {
        "operationId": "get_publisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Publisher id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get publisher by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Publisher not found"
          }
        }
      },
      "put": {
        "operationId": "update_publisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Publisher id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublisherUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Publisher updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Publisher not found"
          }
        }
      },
      "delete": {
        "operationId": "delete_publisher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Publisher id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Publisher deleted successfully"
          },
          "404": {
            "description": "Publisher not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookCreateDto": {
        "type": "object",
        "required": [
          "title",
          "author",
          "publisher_id"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "publisher_id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "BookResponseDto": {
        "type": "object",
        "required": [
          "id",
          "title",
          "author",
          "publisher_id"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "publisher_id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "BookUpdateDto": {
        "type": "object",
        "required": [
          "id",
          "title",
          "author",
          "publisher_id"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "publisher_id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PublisherCreateDto": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "PublisherResponseDto": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PublisherUpdateDto": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}